+++ oneview_redfish_toolkit/tests/api/test_computer_system.py
@@ -18,6 +18,7 @@ import json
 from jsonschema.exceptions import ValidationError
 
 from oneview_redfish_toolkit.api.computer_system import ComputerSystem
+from oneview_redfish_toolkit.api.errors import OneViewRedfishError
 from oneview_redfish_toolkit.tests.base_test import BaseTest
 
 
@@ -39,48 +40,39 @@ class TestComputerSystem(BaseTest):
         ) as f:
             self.server_hardware_types = json.load(f)
 
+        # Loading ServerProfile mockup value
+        with open(
+                'oneview_redfish_toolkit/mockups/oneview/ServerProfile.json'
+        ) as f:
+            self.server_profile = json.load(f)
+
         # Loading ComputerSystem mockup result
         with open(
             'oneview_redfish_toolkit/mockups/redfish/ComputerSystem.json'
         ) as f:
             self.computer_system_mockup = json.load(f)
 
-    def test_class_instantiation(self):
-        # Tests if class is correctly instantiated and validated
-
-        try:
-            computer_system = ComputerSystem(
-                self.server_hardware,
-                self.server_hardware_types
-            )
-        except Exception as e:
-            self.fail("Failed to instantiate ComputerSystem class."
-                      " Error: {}".format(e))
-        self.assertIsInstance(computer_system, ComputerSystem)
-
-    def test_failing_class_instantiation(self):
-        # Tests if validation fail
-        # The name must be a string
-        self.server_hardware["name"] = 1
-
-        with self.assertRaises(ValidationError):
-            ComputerSystem(self.server_hardware, self.server_hardware_types)
-
     def test_serialize(self):
         # Tests the serialize function result against known result
+        computer_system = ComputerSystem(
+            self.server_hardware,
+            self.server_hardware_types,
+            self.server_profile
+        )
 
-        try:
-            computer_system = ComputerSystem(
-                self.server_hardware,
-                self.server_hardware_types
-            )
-        except Exception as e:
-            self.fail("Failed to instantiate ComputerSystem class."
-                      " Error: {}".format(e))
-
-        try:
-            result = json.loads(computer_system.serialize())
-        except Exception as e:
-            self.fail("Failed to serialize. Error: ".format(e))
+        result = json.loads(computer_system.serialize())
 
         self.assertEqual(self.computer_system_mockup, result)
+
+    def test_get_oneview_power_configuration(self):
+        # Tests invalid mapping values of power state
+        #
+        obj = ComputerSystem(self.server_hardware,
+                             self.server_hardware_types,
+                             self.server_profile)
+
+        self.assertRaises(OneViewRedfishError, obj.
+                          get_oneview_power_configuration, "ForceOn")
+
+        self.assertRaises(OneViewRedfishError, obj.
+                          get_oneview_power_configuration, "INVALID")